--- uetemm_notas_back/src/main/java/com/irojas/demojwt/Auth/LoginRequest.java
@@ -1,15 +1,0 @@
-package com.irojas.demojwt.Auth;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class LoginRequest {
-    String username;
-    String password; 
-}
