--- uetemm_notas_back/src/main/java/com/irojas/demojwt/Auth/RegisterRequest.java
@@ -1,18 +1,0 @@
-package com.irojas.demojwt.Auth;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class RegisterRequest {
-    String username;
-    String password;
-    String firstname;
-    String lastname;
-    String country; 
-}
