--- uetemm_notas_back/src/main/java/com/irojas/demojwt/DemoJwtApplication.java
@@ -1,27 +1,0 @@
-package com.irojas.demojwt;
-
-import java.security.SecureRandom;
-import java.util.Base64;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DemoJwtApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DemoJwtApplication.class, args);
-
-		System.out.println("App Running");
-
-	/* 	SecureRandom secureRandom = new SecureRandom();
-
-        byte[] secretKeyBytes = new byte[32];
-        secureRandom.nextBytes(secretKeyBytes);
-
-        String SECRET_KEY = Base64.getEncoder().encodeToString(secretKeyBytes);
-
-        System.out.println(SECRET_KEY);  */
-	}
-
-}
