--- uetemm_notas_back/src/main/java/com/irojas/demojwt/ExeptionHandler/GlobalExeptionHandler.java
@@ -1,21 +1,0 @@
-package com.irojas.demojwt.ExeptionHandler;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class GlobalExeptionHandler {
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<String> handlerArgumentExeption(IllegalArgumentException ex){
-        return new ResponseEntity<String>(ex.getMessage(), HttpStatus.BAD_REQUEST);
-    }
-    
-    @ExceptionHandler(RuntimeException.class)
-    public ResponseEntity<String> handlerRuntimeExeption(RuntimeException ex) {
-        return new ResponseEntity<String>(ex.getMessage(), HttpStatus.BAD_REQUEST);
-    }
-
-}
