--- uetemm_notas_back/src/main/java/com/irojas/demojwt/model/Role.java
@@ -1,7 +1,0 @@
-package com.irojas.demojwt.model;
-
-public enum Role {
-    ADMIN,
-    USER  
-    
-}
