--- uetemm_notas_back/src/main/java/com/irojas/demojwt/repository/UserRepository.java
@@ -1,24 +1,0 @@
-package com.irojas.demojwt.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-import com.irojas.demojwt.model.User;
-
-public interface UserRepository extends JpaRepository<User,Integer> {
-    Optional<User> findByUsername(String username); 
-
-        @Modifying()
-    @Query("update User u set u.firstname=:firstname, u.lastname=:lastname, u.country=:country where u.id = :id")
-    void updateUser(@Param(value = "id") Integer id,   @Param(value = "firstname") String firstname, @Param(value = "lastname") String lastname , @Param(value = "country") String country);
-       
-    @Modifying()
-    @Query("update User u set u.password=:password where u.id = :id") 
-    void updatePassword(@Param(value = "id") Integer id,   @Param(value = "password") String password);
-
-
-}
