--- uetemm_notas_back/src/main/java/com/irojas/demojwt/request/PasswordRequest.java
@@ -1,15 +1,0 @@
-package com.irojas.demojwt.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class PasswordRequest {
-    public int id;
-    public String password;
-}
