--- uetemm_notas_back/src/main/java/com/irojas/demojwt/response/MessageResponse.java
@@ -1,14 +1,0 @@
-package com.irojas.demojwt.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class MessageResponse {
-    String message;
-}
