+++ uetemm_notas_back/src/main/java/com/notas/backend/BackEndApplication.java
@@ -1,0 +1,27 @@
+package com.notas.backend;
+
+import java.security.SecureRandom;
+import java.util.Base64;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BackEndApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BackEndApplication.class, args);
+
+		System.out.println("BackEnd Application Running");
+
+	/* 	SecureRandom secureRandom = new SecureRandom();
+
+        byte[] secretKeyBytes = new byte[32];
+        secureRandom.nextBytes(secretKeyBytes);
+
+        String SECRET_KEY = Base64.getEncoder().encodeToString(secretKeyBytes);
+
+        System.out.println(SECRET_KEY);  */
+	}
+
+}
