+++ uetemm_notas_back/src/main/java/com/notas/backend/auth/AuthResponse.java
@@ -1,0 +1,14 @@
+package com.notas.backend.auth;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class AuthResponse {
+    String token; 
+}
