+++ uetemm_notas_back/src/main/java/com/notas/backend/auth/LoginRequest.java
@@ -1,0 +1,15 @@
+package com.notas.backend.auth;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class LoginRequest {
+    String username;
+    String password; 
+}
