+++ uetemm_notas_back/src/main/java/com/notas/backend/auth/RegisterRequest.java
@@ -1,0 +1,18 @@
+package com.notas.backend.auth;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class RegisterRequest {
+    String username;
+    String password;
+    String firstname;
+    String lastname;
+    String country; 
+}
