+++ uetemm_notas_back/src/main/java/com/notas/backend/dto/CatalogoDTO.java
@@ -1,0 +1,13 @@
+package com.notas.backend.dto;
+
+import lombok.Data;
+
+@Data
+public class CatalogoDTO {
+
+    private int id;
+    private String nombre;
+    private int catalogoPartent;
+
+
+}
