+++ uetemm_notas_back/src/main/java/com/notas/backend/jwt/JwtService.java
@@ -1,0 +1,87 @@
+package com.notas.backend.jwt;
+
+import java.security.Key;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.function.Function;
+
+import javax.crypto.SecretKey;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Service;
+
+import com.notas.backend.model.User;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.security.Keys;
+
+@Service
+public class JwtService {
+
+    @Value("${jwt.secret}")
+    String SECRET_KEY;
+
+    public String getToken(User user) {
+        return getToken(new HashMap<>(), user);
+    }
+
+    private String getToken(Map<String,Object> extraClaims, User user) {
+        return Jwts
+            .builder()
+            .claims(extraClaims)
+            .claim("userId", user.getId())
+            .claim("firstName", user.getFirstname())
+            .claim("lastName", user.getLastname())
+            .subject(user.getUsername())
+            .issuedAt(new Date(System.currentTimeMillis()))
+            .expiration(new Date(System.currentTimeMillis()+1000*60*24))
+            .signWith(getKey())
+            .compact();
+    }
+
+    private SecretKey getKey() {
+       byte[] keyBytes=Decoders.BASE64.decode(SECRET_KEY);
+       return Keys.hmacShaKeyFor(keyBytes);
+    }
+
+    public String getUsernameFromToken(String token) {
+        return getClaim(token, Claims::getSubject);
+    }
+
+    public boolean isTokenValid(String token, UserDetails userDetails) {
+        final String username=getUsernameFromToken(token);
+        return (username.equals(userDetails.getUsername())&& !isTokenExpired(token));
+    }
+
+    private Claims getAllClaims(String token)
+    {
+        return Jwts
+            .parser()
+            .verifyWith(getKey())
+            .build()
+            .parseSignedClaims(token)
+            .getPayload();
+    }
+
+    public <T> T getClaim(String token, Function<Claims,T> claimsResolver)
+    {
+        final Claims claims=getAllClaims(token);
+        return claimsResolver.apply(claims);
+    }
+
+    private Date getExpiration(String token)
+    {
+        return getClaim(token, Claims::getExpiration);
+    }
+
+    private boolean isTokenExpired(String token)
+    {
+        return getExpiration(token).before(new Date());
+    }
+    
+}
