+++ uetemm_notas_back/src/main/java/com/notas/backend/model/Role.java
@@ -1,0 +1,7 @@
+package com.notas.backend.model;
+
+public enum Role {
+    ADMIN,
+    USER  
+    
+}
