+++ uetemm_notas_back/src/main/java/com/notas/backend/request/PasswordRequest.java
@@ -1,0 +1,15 @@
+package com.notas.backend.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class PasswordRequest {
+    public int id;
+    public String password;
+}
