+++ uetemm_notas_back/src/main/java/com/notas/backend/request/UserRequest.java
@@ -1,0 +1,19 @@
+package com.notas.backend.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserRequest {
+    public int id;
+    public String username;
+    public String firstname;
+    public String lastname;
+    public String country;
+    public int estadoCivil;
+}
