+++ uetemm_notas_back/src/main/java/com/notas/backend/response/MessageResponse.java
@@ -1,0 +1,14 @@
+package com.notas.backend.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class MessageResponse {
+    String message;
+}
