+++ uetemm_notas_front/uetemm_app/src/app/pages/user-details/user-details.component.ts
@@ -4,6 +4,8 @@ import { FormBuilder, Validators } from '@angular/forms';
 import { UserService } from '../../services/user/user.service';
 import { environment } from '../../../environments/environment.development';
 import { LoginService } from '../../services/auth/login.service';
+import { Catalogo } from '../../services/catalogo/catalogo';
+import { CatalogoService } from '../../services/catalogo/catalogo.service';
 
 @Component({
   selector: 'app-user-details',
@@ -15,26 +17,35 @@ export class UserDetailsComponent {
   user?: User;
   userLoggedOn: boolean = false;
   editMode: boolean = false;
+  
+  catalogoEstadoCivil: Catalogo[] = [];
+  selectedEstadoCivil: number = 0;
+  
 
-  registerForm = this.formBuilder.group({
+
+  userDetailsForm = this.formBuilder.group({
     id: [''],
-    lastname: ['', Validators.required],
     firstname: ['', Validators.required],
+    lastname: ['', Validators.required],
+    username: ['', Validators.required],
     country: ['', Validators.required],
+    estadoCivil: ['', Validators.required],
   });
 
   constructor(
     private userService: UserService,
     private formBuilder: FormBuilder,
-    private loginService: LoginService
+    private loginService: LoginService,
+    private catalogoService: CatalogoService
   ) {
     this.userService.getUser(this.loginService.userToken).subscribe({
       next: (userData) => {
         this.user = userData;
-        this.registerForm.controls.id.setValue(userData.id.toString());
-        this.registerForm.controls.firstname.setValue(userData.firstname);
-        this.registerForm.controls.lastname.setValue(userData.lastname);
-        this.registerForm.controls.country.setValue(userData.country);
+        this.userDetailsForm.controls.id.setValue(userData.id.toString());
+        this.userDetailsForm.controls.firstname.setValue(userData.firstname);
+        this.userDetailsForm.controls.lastname.setValue(userData.lastname);
+        this.userDetailsForm.controls.username.setValue(userData.username);
+        this.userDetailsForm.controls.country.setValue(userData.country);
       },
       error: (errorData) => {
         this.errorMessage = errorData;
@@ -49,29 +60,38 @@ export class UserDetailsComponent {
         this.userLoggedOn = userLoggedOn;
       },
     });
+
+    this.catalogoService.getEstadoCivilLista().subscribe(
+      (data: Catalogo[])=>{
+        this.catalogoEstadoCivil = data;
+        console.log(this.catalogoEstadoCivil)
+      },
+      (error) => {
+        console.error('Error fetching catalogos', error);
+      }
+    );
   }
 
   get firstname() {
-    return this.registerForm.controls.firstname;
+    return this.userDetailsForm.controls.firstname;
   }
 
   get lastname() {
-    return this.registerForm.controls.lastname;
+    return this.userDetailsForm.controls.lastname;
   }
 
   get country() {
-    return this.registerForm.controls.country;
+    return this.userDetailsForm.controls.country;
   }
 
   saveUserDetailsData() {
-    if (this.registerForm.valid) {
-      console.log(this.registerForm);
+    if (this.userDetailsForm.valid) {
       this.userService
-        .updateUser(this.registerForm.value as unknown as User)
+        .updateUser(this.userDetailsForm.value as unknown as User)
         .subscribe({
           next: () => {
             this.editMode = false;
-            this.user = this.registerForm.value as unknown as User;
+            this.user = this.userDetailsForm.value as unknown as User;
           },
           error: (errorData) => console.error(errorData),
         });
