+++ uetemm_notas_front/uetemm_app/src/app/services/auth/user.ts
@@ -7,5 +7,6 @@ export interface User{
     role?: string;
     mensaje?: string;
     country:string;
+    estadoCivil:number;
 
 }
