+++ uetemm_notas_front/uetemm_app/src/app/services/catalogo/catalogo.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { CatalogoService } from './catalogo.service';
+
+describe('CatalogoService', () => {
+  let service: CatalogoService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(CatalogoService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
