+++ uetemm_notas_front/uetemm_app/src/app/services/catalogo/catalogo.ts
@@ -1,0 +1,5 @@
+export interface Catalogo {
+    id: number;
+    nombre: string;
+    catalogoParent: number;
+  }
