+++ uetemm_notas_front/uetemm_app/src/environments/environment.development.ts
@@ -1,7 +1,7 @@
 export const environment = {
     production: false,
     urlHost: 'http://localhost:8091',
-    urlApi: 'http://localhost:8091/api/v1',
+    urlApi: 'http://localhost:8091/api/',
